Move remote DHT logic into context
